Refactor by reusing IntervalLooper
