+++ app/src/main/res/layout/video_dialog.xml
@@ -9,7 +9,7 @@
     <VideoView
         android:id="@+id/videoView_recordingVideo_videoDialog"
         android:layout_width="match_parent"
-        android:layout_height="0dp"
+        android:layout_height="wrap_content"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
